Revision of the first few pages (#41)
* Update overview.md

* Update overview.md

language fix

* Update overview.md

* Update pre_annotations.md

Language fix

* Update kognic_auth.md

Improve language and the structure of the description to be more clear and precise.

* Update key_concepts.md

Improved the language to be more precise and clear. Also delete unnecessary information to make it shorter

* Update key_concepts.md

correcting a minor typo

* Update project.md

Improve the language and remove some long and repeated sentences that are explained elsewhere.

* Update overview.md

Read until Frame. Will continue after the new year.

* read till resources

* Update docs-src/docs/kognic-io/project.md

language

Co-authored-by: Oskar Olin <oskar.olin@annotell.com>

* Update docs-src/docs/kognic-io/project.md

language fix

Co-authored-by: Oskar Olin <oskar.olin@annotell.com>

* Deleted more repetitive sentences after Oscar's comments.

---------

Co-authored-by: Oskar Olin <oskar.olin@annotell.com>